Code-for-MuChin-AP(AnnotationPlatform)/backend_api/requirements.txt
Flask-SQLAlchemy == 3.1.1
pymysql == 1.1.0
cryptography == 41.0.4
flask == 2.3.3
PyJWT == 2.8.0
pypinyin == 0.49.0
zhconv == 1.4.3
gunicorn == 21.2.0
flask-cors == 4.0.0
python-dotenv == 1.0.0
openai == 1.3.3
